notifier hooks for option name/value manager

+++ admin/options_values_manager.php
@@ -461,14 +461,19 @@ die('I SEE match from products_id:' . $copy_from_products_id . ' options_id_from
             $downloads_remove_query = "select products_attributes_id from " . TABLE_PRODUCTS_ATTRIBUTES . " where products_id='" . $current_products_id . "' and options_id='" . $options_id_from . "' and options_values_id='" . $options_values_values_id_from . "'";
             $downloads_remove = $db->Execute($downloads_remove_query);
 
+            $remove_downloads_ids = array();
+            foreach($downloads_remove as $row) {
+              $remove_downloads_ids[] = $row['products_attributes_id'];
+            }
+            $zco_notifier->notify('OPTIONS_VALUES_MANAGER_DELETE_VALUES_OF_OPTIONNAME', array('current_products_id' => $current_products_id, 'remove_ids' => $remove_downloads_ids, 'options_id'=>$options_id_from, 'options_values_id'=>$options_values_values_id_from));
+
             $sql = "delete from " . TABLE_PRODUCTS_ATTRIBUTES . " where products_id='" . $current_products_id . "' and options_id='" . $options_id_from . "' and options_values_id='" . $options_values_values_id_from . "'";
             $delete_selected = $db->Execute($sql);
 
             // delete associated downloads
-            while (!$downloads_remove->EOF) {
+            if (sizeof($remove_downloads_ids)) {
               $db->Execute("delete from " . TABLE_PRODUCTS_ATTRIBUTES_DOWNLOAD . "
-                            where products_attributes_id='" . $downloads_remove->fields['products_attributes_id'] . "'");
-              $downloads_remove->MoveNext();
+                            where products_attributes_id in (" . implode(',', $remove_downloads_ids) . ")");
             }
             // count deleted attribute
             $new_attribute++;
